+++ webapp/assets/js/usuario.tsx
@@ -1,3 +1,5 @@
+import Swal, { SweetAlertResult } from "sweetalert2";
+
 $('#para-de-seguir').on('click', pararDeSeguir);
 $('#seguir').on('click', seguir);
 $('#ditar-usuario').on('submit', editar);
@@ -5,37 +7,37 @@ $('#atualizar-senha').on('submit', atualizarSenha)
 $('#delete-usuario').on('click', deletarUsuario);
 
 
-function pararDeSeguir() {
+function pararDeSeguir(this: any) {
     const usuarioId = $(this).data('usuario-id');
     $(this).prop('disabled', true);
 
     $.ajax({
         url: `usuarios/${usuarioId}/parar-de-seguir`,
         method: "POST"
-    }).done(function() {
-        window.location = `/usuarios/${usuarioId}`;
-    }).fail(function() {
+    }).done(function () {
+        window.location.replace(`/usuarios/${usuarioId}`);
+    }).fail(function () {
         Swal.fire("Ops...", "Erro ao parar de seguir o usuário!", "error");
         $('#parar-de-seguir').prop('disabled', false);
     });
 }
 
-function seguir() {
+function seguir(this: any) {
     const usuarioId = $(this).data('usuario-id');
     $(this).prop('disabled', true);
 
     $.ajax({
         url: `usuarios/${usuarioId}/seguir`,
         method: "POST"
-    }).done(function() {
-        window.location = `/usuarios/${usuarioId}`;
-    }).fail(function() {
+    }).done(function () {
+        window.location.replace(`/usuarios/${usuarioId}`);
+    }).fail(function () {
         Swal.fire("Ops...", "Erro ao seguir o usuário!", "error");
         $('#seguir').prop('disabled', false);
     });
 }
 
-function editar(evento) {
+function editar(evento: { preventDefault: () => void; }) {
     evento.preventDefault();
 
     $.ajax({
@@ -46,25 +48,25 @@ function editar(evento) {
             email: $('#email').val(),
             nick: $('#nick').val(),
         }
-    }).done(function() {
+    }).done(function () {
         Swal.fire("Sucesso!", "Usuário atualizado com sucesso!", "success")
-        .then(function() {
-            window.location = "/perfil";
-        });
-    }).fail(function() {
+            .then(function () {
+                window.location.replace("/perfil");
+            });
+    }).fail(function () {
         Swal.fire("Ops...", "Erro ao atualizar o usuário", "error");
     });
 }
 
-function atualizarSenha(evento) {
+function atualizarSenha(evento: { preventDefault: () => void; }) {
     evento.preventDefault();
 
-    if ($('#nova-senha'),val() != $('#confirmar-senha').val()) {
+    if ($('#nova-senha').val() != $('#confirmar-senha').val()) {
         Swal.fire("Ops...", "As senhas devem ser iguais!", "warning");
         return;
     }
 
-    $.ajax ({
+    $.ajax({
         url: "/atualizar-senha",
         method: "POST",
         data: {
@@ -73,10 +75,10 @@ function atualizarSenha(evento) {
         }
     }).done(function () {
         Swal.fire("Sucesso!", "A senha foi atualizada com sucesso!", "success")
-        .then(function() {
-            window.location = "/perfil";
+            .then(function () {
+                window.location.replace("/perfil");
             })
-    }).fail(function() {
+    }).fail(function () {
         Swal.fire("Ops...", "Erro ao atualizar a senha!", "error");
     });
 }
@@ -88,17 +90,17 @@ function deletarUsuario() {
         showCancelButton: true,
         cancelButtonText: "Cancelar",
         icon: "warning"
-    }).then(function(confirmacao) {
+    }).then((confirmacao: SweetAlertResult<any>) => {
         if (confirmacao.value) {
             $.ajax({
                 url: "/deletar-usuario",
                 method: "DELETE"
-            }).done(function() {
+            }).done(function () {
                 Swal.fire("Sucesso!", "Seu usuário foi excluído com sucesso!", "success")
-                .then(function() {
-                    window.location = "/logout";
-                })
-            }).fail(function() {
+                    .then(function () {
+                        window.location.replace("/logout");
+                    })
+            }).fail(function () {
                 Swal.fire("Ops...", "Ocorreu um erro ao excluir o seu usuário", "error");
             });
         }
